requirements.txt
telebot==0.0.5
psycopg2-binary==2.9.6
python-dotenv==1.0.0
psycopg2~=2.9.10
yookassa~=3.5.0
Flask~=3.1.1
requests~=2.32.4